Replace individual measurement unit options with Metric or Imperial
This will allow us to be smarter about the choice of exact scale. For example, snow will be shown in inches and cm or feet and metres if it's deep enough. Rain will be shown in mm or inches. Visibility will be shown in metres and kilometres or feet and miles.